+++ CHANGELOG.md
@@ -18,7 +18,7 @@ at https://github.com/emperorshammer/issues-incoming-queue.
 
 ## ------------------------------------------------------------------------------------------------##
 
-## [v3.0.4] 2023-02-18
+## [v3.0.4] 2023-04-16
 
 ### Added
 - Support for BGCOM position with appropriate collar insignia. (@6958/@238)
@@ -29,6 +29,7 @@ at https://github.com/emperorshammer/issues-incoming-queue.
 - Updated ISM to new TC logo design. (@238)
 - Updated MoS for full implementation. (@238)
 - Corrected MoC device to proper oak leaf cluster. (@238)
+- Group Commanding Officer (GCO) has been changed back to Subgroup Commander (SGCOM). (@238)
 
 ## ------------------------------------------------------------------------------------------------##